refactor(Button): change reading type from int to uint8_t for improved memory usage

+++ src/traffic_light/Button.cpp
@@ -18,7 +18,7 @@ void Button::begin() {
 }
 
 void Button::update() {
-  int reading = digitalRead(pin);
+  uint8_t reading = digitalRead(pin);
 
   // Reset debounce timer on state change
   if (reading != lastReading) {
